Added gradient finder for routes

+++ src/data_types/gc/effort.rs
@@ -10,7 +10,7 @@ pub struct Effort {
     pub activity_id: DocumentId,
     pub moving_time: i32,
     
-    pub distance_from_start: i32
+    pub distance_from_start: f32
 }
 
 impl Identifiable for Effort {
